refactor: moved data classes to tables package
